app_ui/requirements.txt
AppKit==0.2.8
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
cffi==1.17.1
chardet==5.2.0
ConfigParser==7.2.0
contextlib2==21.6.0
crypto==1.4.1
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
evdev==1.9.1
fido2==1.2.0
filelock==3.18.0
gssapi==1.9.0
h2==4.2.0
HTMLParser==0.0.2
importlib_resources==6.5.2
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
js==1.0
keyring==25.6.0
matplotlib==3.10.1
mock==5.2.0
mysql_connector_python==9.2.0
nio==3.4.2
numpy==2.2.4
oci==2.149.0
olefile==0.47
pandas==2.2.3
protobuf==6.30.1
pyasn1==0.6.1
pycryptodome_test_vectors==1.0.21
pyodide==0.0.2
pyOpenSSL==25.0.0
PyQt6==6.8.1
PySide6==6.8.2.1
pytest==8.3.5
pywin32==310
Quartz==0.0.1.dev0
redis==5.2.1
sets==0.3.2
setuptools==75.8.0
setuptools_scm==8.2.0
simplejson==3.20.1
slack_sdk==3.35.0
sqlparse==0.5.3
sspi==0.1.0
tensorflow==2.19.0
thread==2.0.5
urllib3_secure_extra==0.1.0
Werkzeug==3.1.3
Xlib==0.21
xmlrpclib==1.0.1
zstandard==0.23.0
